+++ libs/cardano-protocol/src/Cardano/Protocol/Leios/BlockHeader.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
@@ -10,6 +9,7 @@
 {-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE UndecidableSuperClasses #-}
 {-# LANGUAGE ViewPatterns #-}
 
 -- | Block header associated with Leios.
@@ -19,10 +19,13 @@
 -- "Cardano.Protocol.Praos.BlockHeader" for details.
 module Cardano.Protocol.Leios.BlockHeader (
   Header (HeaderConstr, Header, headerBody, headerSig),
+  bodyHeaderL,
+  sigHeaderL,
   HeaderBody (..),
   EbReferencesAnnouncement (..),
   headerHash,
   headerSize,
+  LeiosEraBlockHeader (..),
 ) where
 
 import qualified Cardano.Crypto.Hash as Hash
@@ -46,7 +49,7 @@ import Cardano.Ledger.Binary (
   unCBORGroup,
  )
 import qualified Cardano.Ledger.Binary.Plain as Plain
-import Cardano.Ledger.Block (Block (..), EraBlockHeader (..))
+import Cardano.Ledger.Block (Block (..), EraBlockHeader (..), headerBlockL)
 import Cardano.Ledger.Core (Era)
 import Cardano.Ledger.Hashes (
   EraIndependentBlockBody,
@@ -79,7 +82,7 @@ import Control.DeepSeq (NFData)
 import Data.Maybe.Strict (StrictMaybe (..))
 import Data.Word (Word32)
 import GHC.Generics (Generic)
-import Lens.Micro (lens, to)
+import Lens.Micro (Lens', lens, to)
 import NoThunks.Class (NoThunks (..))
 
 -- | Announcement of an Endorser Block (EB).
@@ -263,32 +266,31 @@ deriving via
   instance
     Crypto c => DecCBOR (Annotator (Header c))
 
+bodyHeaderL :: Crypto c => Lens' (Header c) (HeaderBody c)
+bodyHeaderL = lens headerBody (\h b -> h {headerBody = b})
+
+sigHeaderL :: Crypto c => Lens' (Header c) (KES.SignedKES (KES c) (HeaderBody c))
+sigHeaderL = lens headerSig (\h s -> h {headerSig = s})
+
 instance (Crypto c, Era era) => EraBlockHeader (Header c) era where
-  blockIssuerBlockHeaderG =
-    to (\(Block (Header hb _) _) -> hashKey (hbVk hb))
   blockHeaderSizeBlockHeaderG =
-    to (\(Block hdr _) -> originalBytesSize hdr)
+    headerBlockL . to originalBytesSize
+  blockIssuerBlockHeaderG =
+    headerBlockL . bodyHeaderL . to (hashKey . hbVk)
   blockBodySizeBlockHeaderL =
-    lens
-      (\(Block (Header hb _) _) -> hbBodySize hb)
-      ( \(Block (Header hb sig) body) sz ->
-          Block (Header hb {hbBodySize = sz} sig) body
-      )
+    headerBlockL . bodyHeaderL . lens hbBodySize (\hb sz -> hb {hbBodySize = sz})
   blockBodyHashBlockHeaderL =
-    lens
-      (\(Block (Header hb _) _) -> hbBodyHash hb)
-      ( \(Block (Header hb sig) body) h ->
-          Block (Header hb {hbBodyHash = h} sig) body
-      )
+    headerBlockL . bodyHeaderL . lens hbBodyHash (\hb h -> hb {hbBodyHash = h})
   slotNoBlockHeaderL =
-    lens
-      (\(Block (Header hb _) _) -> hbSlotNo hb)
-      ( \(Block (Header hb sig) body) s ->
-          Block (Header hb {hbSlotNo = s} sig) body
-      )
+    headerBlockL . bodyHeaderL . lens hbSlotNo (\hb sn -> hb {hbSlotNo = sn})
   protVerBlockHeaderL =
-    lens
-      (\(Block (Header hb _) _) -> hbProtVer hb)
-      ( \(Block (Header hb sig) body) pv ->
-          Block (Header hb {hbProtVer = pv} sig) body
-      )
+    headerBlockL . bodyHeaderL . lens hbProtVer (\hb pv -> hb {hbProtVer = pv})
+
+class EraBlockHeader h era => LeiosEraBlockHeader h era where
+  ebReferencesAnnouncementBlockHeaderL :: Lens' (Block h era) (StrictMaybe EbReferencesAnnouncement)
+
+instance (Crypto c, Era era) => LeiosEraBlockHeader (Header c) era where
+  ebReferencesAnnouncementBlockHeaderL =
+    headerBlockL
+      . bodyHeaderL
+      . lens hbEbReferencesAnnouncement (\hb ma -> hb {hbEbReferencesAnnouncement = ma})